# To fix extra memory allocation when using circular padding