--- a/torch/nn/functional.py
+++ b/torch/nn/functional.py
@@ -3876,37 +3876,55 @@ def _pad_circular(input, padding):
         assert padding[-(a*2+2)] <= size
 
     # Get shape of padded array
-    new_shape = shape[:2]
+    padded_shape = shape[:2]
     for a, size in enumerate(shape[2:]):
-        new_shape += (size + padding[-(a*2+1)] + padding[-(a*2+2)],)
+        padded_shape += (size + padding[-(a*2+1)] + padding[-(a*2+2)],)
 
-    out = torch.empty(new_shape, dtype=input.dtype, layout=input.layout,
+    out = torch.empty(padded_shape, dtype=input.dtype, layout=input.layout,
                       device=input.device)
 
     # Put original array in padded array
     if ndim == 1:
-        out[..., padding[-2]:-padding[-1]] = input
+        d0 = padding[-2]
+        d1 = padded_shape[2] - padding[-1]
+        out[..., d0:d1] = input
     elif ndim == 2:
-        out[..., padding[-2]:-padding[-1], padding[-4]:-padding[-3]] = input
+        d0 = padding[-2]
+        d1 = padded_shape[2] - padding[-1]
+        h0 = padding[-4]
+        h1 = padded_shape[3] - padding[-3]
+        out[..., d0:d1, h0:h1] = input
     elif ndim == 3:
-        out[..., padding[-2]:-padding[-1], padding[-4]:-padding[-3], padding[-6]:-padding[-5]] = input
+        d0 = padding[-2]
+        d1 = padded_shape[2] - padding[-1]
+        h0 = padding[-4]
+        h1 = padded_shape[3] - padding[-3]
+        w0 = padding[-6]
+        w1 = padded_shape[4] - padding[-5]
+        out[..., d0:d1, h0:h1, w0:w1] = input
 
-    # Pad right side, then left side.
-    # Corners will be written more than once when ndim > 1
+    # The following steps first pad the left side, then pad the right side
+    # Note: Corners will be written more than once when ndim > 1
 
-    # Pad first conv dim
-    out[:, :, :padding[-2]] = out[:, :, -(padding[-2] + padding[-1]):-padding[-1]]
-    out[:, :, -padding[-1]:] = out[:, :, padding[-2]:(padding[-2] + padding[-1])]
+    # Pad first conv dim (depth-wise)
+    out[:, :, :padding[-2]] = \
+        out[:, :, -(padding[-2] + padding[-1]):-padding[-1]]
+    out[:, :, (padded_shape[2]-padding[-1]):] = \
+        out[:, :, padding[-2]:(padding[-2] + padding[-1])]
 
     if len(padding) > 2:
-        # Pad second conv dim
-        out[:, :, :, :padding[-4]] = out[:, :, :, -(padding[-4] + padding[-3]):-padding[-3]]
-        out[:, :, :, -padding[-3]:] = out[:, :, :, padding[-4]:(padding[-4] + padding[-3])]
+        # Pad second conv dim (height-wise)
+        out[:, :, :, :padding[-4]] = \
+            out[:, :, :, -(padding[-4] + padding[-3]):-padding[-3]]
+        out[:, :, :, (padded_shape[3]-padding[-3]):] = \
+            out[:, :, :, padding[-4]:(padding[-4] + padding[-3])]
 
     if len(padding) > 4:
-        # Pad third conv dim
-        out[:, :, :, :, :padding[-6]] = out[:, :, :, :, -(padding[-6] + padding[-5]):-padding[-5]]
-        out[:, :, :, :, -padding[-5]:] = out[:, :, :, :, padding[-6]:(padding[-6] + padding[-5])]
+        # Pad third conv dim (width-wise)
+        out[:, :, :, :, :padding[-6]] = \
+            out[:, :, :, :, -(padding[-6] + padding[-5]):-padding[-5]]
+        out[:, :, :, :, (padded_shape[4]-padding[-5]):] = \
+            out[:, :, :, :, padding[-6]:(padding[-6] + padding[-5])]
 
     return out