--- a/torch/nn/functional.py
+++ b/torch/nn/functional.py
@@ -3929,37 +3929,76 @@ def _pad_circular(input, padding):
         out[..., out_d[0]:out_d[1], out_h[0]:out_h[1], out_w[0]:out_w[1]] = \
             input[..., in_d[0]:in_d[1], in_h[0]:in_h[1], in_w[0]:in_w[1]]
 
-    # The following steps first pad the left side, then pad the right side
-    # Note: Corners will be written more than once when ndim > 1
+    # The following steps first pad the beginning of the tensor (left side),
+    # and then pad the end of the tensor (right side).
+    # Note: Corners will be written more than once when ndim > 1.
 
     # Only in cases where padding values are > 0 are when additional copying
-    # is required
+    # is required.
 
-    # Pad first conv dim (depth-wise)
+    def calc_pad_begin(size, padding):
+        """Calculates indices for padding the beginning of the tensor.
+
+        Args:
+            size (int): Size of dimension of padded tensor.
+            padding (tuple): Length 2 tuple containing padding values for
+                beginning and end of tensor.
+
+        Returns:
+            src_idx (tuple): Start and end indices for slicing source.
+            dest_idx (tuple): Start and end indices for slicing destination.
+        """
+        padding1 = max(padding[1], 0)
+        src_idx = (size - padding[0] - padding1, size - padding1)
+        dest_idx = (0, padding[0])
+        return src_idx, dest_idx
+
+    def calc_pad_end(size, padding):
+        """Calculates indices for padding the end of the tensor.
+
+        Args:
+            size (int): Size of dimension of padded tensor.
+            padding (tuple): Length 2 tuple containing padding values for
+                beginning and end of tensor.
+
+        Returns:
+            src_idx (tuple): Start and end indices for slicing source.
+            dest_idx (tuple): Start and end indices for slicing destination.
+        """
+        padding0 = max(padding[0], 0)
+        src_idx = (padding0, padding0 + padding[1])
+        dest_idx = (size - padding[1], size)
+        return src_idx, dest_idx
+
+    # Pad first dimension (depth)
     if padding[-2] > 0:
-        out[:, :, :padding[-2]] = \
-            out[:, :, -(padding[-2] + padding[-1]):-padding[-1]]
+        src_idx, dest_idx = calc_pad_begin(out_shape[2], padding[-2:])
+        out[:, :, dest_idx[0]:dest_idx[1]] = out[:, :, src_idx[0]:src_idx[1]]
     if padding[-1] > 0:
-        out[:, :, (out_shape[2] - padding[-1]):] = \
-            out[:, :, padding[-2]:(padding[-2] + padding[-1])]
+        src_idx, dest_idx = calc_pad_end(out_shape[2], padding[-2:])
+        out[:, :, dest_idx[0]:dest_idx[1]] = out[:, :, src_idx[0]:src_idx[1]]
 
-    # Pad second conv dim (height-wise)
+    # Pad second dimension (height)
     if len(padding) > 2:
         if padding[-4] > 0:
-            out[:, :, :, :padding[-4]] = \
-                out[:, :, :, -(padding[-4] + padding[-3]):-padding[-3]]
+            src_idx, dest_idx = calc_pad_begin(out_shape[3], padding[-4:-2])
+            out[:, :, :, dest_idx[0]:dest_idx[1]] = \
+                out[:, :, :, src_idx[0]:src_idx[1]]
         if padding[-3] > 0:
-            out[:, :, :, (out_shape[3] - padding[-3]):] = \
-                out[:, :, :, padding[-4]:(padding[-4] + padding[-3])]
+            src_idx, dest_idx = calc_pad_end(out_shape[3], padding[-4:-2])
+            out[:, :, :, dest_idx[0]:dest_idx[1]] = \
+                out[:, :, :, src_idx[0]:src_idx[1]]
 
-    # Pad third conv dim (width-wise)
+    # Pad third dimension (width)
     if len(padding) > 4:
         if padding[-6] > 0:
-            out[:, :, :, :, :padding[-6]] = \
-                out[:, :, :, :, -(padding[-6] + padding[-5]):-padding[-5]]
+            src_idx, dest_idx = calc_pad_begin(out_shape[4], padding[-6:-4])
+            out[:, :, :, :, dest_idx[0]:dest_idx[1]] = \
+                out[:, :, :, :, src_idx[0]:src_idx[1]]
         if padding[-5] > 0:
-            out[:, :, :, :, (out_shape[4] - padding[-5]):] = \
-                out[:, :, :, :, padding[-6]:(padding[-6] + padding[-5])]
+            src_idx, dest_idx = calc_pad_end(out_shape[4], padding[-6:-4])
+            out[:, :, :, :, dest_idx[0]:dest_idx[1]] = \
+                out[:, :, :, :, src_idx[0]:src_idx[1]]
 
     return out