**BC-breaking Note:**

This PR changes the behavior of circular padding when using negative or zero values. Previously these cases would not work properly, with negative circular padding not reducing the sizes of the input tensor properly, and with zero circular padding sometimes causing dimensions not to be padded at all. This PR corrects circular padding's behavior. 

**Original PR Summary**

For fixing https://github.com/pytorch/pytorch/issues/39256