@chcodes 2020-06-02
It's failing when the value for padding is (0, 0, -1, -1).
Does anyone know what the intended behaviour is when padding is -1?

The current circular padding does this:
```python
import torch
import torch.nn.functional as F


inputs = torch.rand(1, 3, 4, 4)
out = F.pad(inputs, (0, 0, -1, -1), mode='circular')
print(out.shape)  # torch.Size([1, 3, 7, 4])
```
It takes 3 slices and pads one side with it, and nothing happens to the other side. Don't think this is right.

cc @peterbell10