@peterbell10 2020-06-02
Well, according to `mode="constant"` the behaviour should be to narrow the tensor so that 
```cpp
out_dim[i] = in_dim[i] + padding[2*i] + padding[2*i + 1]
```

https://github.com/pytorch/pytorch/blob/de5b8797e675d8bb5867d9e06dd08e0217db07db/aten/src/ATen/native/ConstantPadNd.cpp#L25-L31