@chcodes 2020-06-02
If I am understanding this correctly, negative values will essentially crop the array?

Do we want the general padding function to be able to do cropping? For all modes: circular, reflection, replication, etc. I don't think users would intuitively look to use F.pad when they want to crop the image.

An argument can be made that the backward function of padding involves cropping, but certainly does more than just cropping in the cases where the padding mode isn't zeros.

If we don't want the general padding function to be able to do cropping, then I think we should remove the test on values of -1. And currently the circular padding doesn't do cropping when receiving negative padding values anyway.