@mruberry 2020-06-03
Hey @chcodes, thanks for submitting this PR!

I appreciate your goal of trying to reduce memory usage by allocating fewer intermediate tensors, but I also worry about the risk of this PR because it's causing many tests to fail, and I don't think we want to change the historic behavior to preserve backwards compatibility. That is, maybe we SHOULD have different padding behavior, but that's a different discussion that will require great care and possibly deprecation warnings. 

Is there any way we can "de-risk" this PR and have high confidence it still implements the historic behavior while reducing total memory usage?