@chcodes 2020-06-08
Ah sorry circular padding was the only mode that didn't narrow the array. Ok I've changed my circular padding to correctly narrow the array.