@mruberry 2020-06-16
Hey @chcodes, thanks for updating this PR, it's great to see it's passing all the tests! 

However I'm still concerned about a few meta issues:

- Our padding documentation is very poor, how do we know this PR implements the same behavior we historically did?
- How do we know this PR actually uses less memory than the previous approach? 

I really appreciate you want to improve this code's documentation and performance. Maybe some local analysis could help address these issues? For example, passing a variety of arguments that test the different code paths and observing they give the same results? And showing what the memory usage is in each case?