@chcodes 2020-07-13
I added timing tests. The new function runs faster on both CPU and GPU on my machine. I just appended the test to the script I was using before.

Will work on the other comments.

```python
import torch
from torch.nn import functional as F
from time import time


# New circular padding function
def _pad_circular(input, padding):
    # type: (Tensor, List[int]) -> Tensor
    """
    Args:
        input: Tensor that follows the formatting of the input to convolution
            layers.
        padding: Tuple with length two times the degree of the convolution. The
            order of the integers in the tuple are shown in the following
            example:
            For 3D convolutions:
                padding[-2] is the amount of padding applied to the beginning
                    of the depth dimension.
                padding[-1] is the amount of padding applied to the end of the
                    depth dimension.
                padding[-4] is the amount of padding applied to the beginning
                    of the height dimension.
                padding[-3] is the amount of padding applied to the end of the
                    height dimension.
                padding[-6] is the amount of padding applied to the beginning
                    of the width dimension.
                padding[-5] is the amount of padding applied to the end of the
                    width dimension.
    Returns:
        out: Tensor with padded shape.
    """
    shape = input.shape
    ndim = len(shape[2:])

    # Only supports wrapping around once
    for a, size in enumerate(shape[2:]):
        assert padding[-(a * 2 + 1)] <= size
        assert padding[-(a * 2 + 2)] <= size

    # Get shape of padded array
    padded_shape = shape[:2]
    for a, size in enumerate(shape[2:]):
        padded_shape += (size + padding[-(a * 2 + 1)] + padding[-(a * 2 + 2)],)

    out = torch.empty(padded_shape, dtype=input.dtype, layout=input.layout,
                      device=input.device)

    # Put original array in padded array
    if ndim == 1:
        out_d0 = max(padding[-2], 0)
        out_d1 = padded_shape[2] - max(padding[-1], 0)

        in_d0 = max(-padding[-2], 0)
        in_d1 = shape[2] - max(-padding[-1], 0)

        out[..., out_d0:out_d1] = input[..., in_d0:in_d1]
    elif ndim == 2:
        out_d0 = max(padding[-2], 0)
        out_d1 = padded_shape[2] - max(padding[-1], 0)

        out_h0 = max(padding[-4], 0)
        out_h1 = padded_shape[3] - max(padding[-3], 0)

        in_d0 = max(-padding[-2], 0)
        in_d1 = shape[2] - max(-padding[-1], 0)

        in_h0 = max(-padding[-4], 0)
        in_h1 = shape[3] - max(-padding[-3], 0)

        out[..., out_d0:out_d1, out_h0:out_h1] = \
            input[..., in_d0:in_d1, in_h0:in_h1]
    elif ndim == 3:
        out_d0 = max(padding[-2], 0)
        out_d1 = padded_shape[2] - max(padding[-1], 0)

        out_h0 = max(padding[-4], 0)
        out_h1 = padded_shape[3] - max(padding[-3], 0)

        out_w0 = max(padding[-6], 0)
        out_w1 = padded_shape[4] - max(padding[-5], 0)

        in_d0 = max(-padding[-2], 0)
        in_d1 = shape[2] - max(-padding[-1], 0)

        in_h0 = max(-padding[-4], 0)
        in_h1 = shape[3] - max(-padding[-3], 0)

        in_w0 = max(-padding[-6], 0)
        in_w1 = shape[4] - max(-padding[-5], 0)

        out[..., out_d0:out_d1, out_h0:out_h1, out_w0:out_w1] = \
            input[..., in_d0:in_d1, in_h0:in_h1, in_w0:in_w1]

    # The following steps first pad the left side, then pad the right side
    # Note: Corners will be written more than once when ndim > 1

    # Only in cases where padding values are > 0 are when additional copying
    # is required

    # Pad first conv dim (depth-wise)
    if padding[-2] > 0:
        out[:, :, :padding[-2]] = \
            out[:, :, -(padding[-2] + padding[-1]):-padding[-1]]
    if padding[-1] > 0:
        out[:, :, (padded_shape[2] - padding[-1]):] = \
            out[:, :, padding[-2]:(padding[-2] + padding[-1])]

    # Pad second conv dim (height-wise)
    if len(padding) > 2:
        if padding[-4] > 0:
            out[:, :, :, :padding[-4]] = \
                out[:, :, :, -(padding[-4] + padding[-3]):-padding[-3]]
        if padding[-3] > 0:
            out[:, :, :, (padded_shape[3] - padding[-3]):] = \
                out[:, :, :, padding[-4]:(padding[-4] + padding[-3])]

    # Pad third conv dim (width-wise)
    if len(padding) > 4:
        if padding[-6] > 0:
            out[:, :, :, :, :padding[-6]] = \
                out[:, :, :, :, -(padding[-6] + padding[-5]):-padding[-5]]
        if padding[-5] > 0:
            out[:, :, :, :, (padded_shape[4] - padding[-5]):] = \
                out[:, :, :, :, padding[-6]:(padding[-6] + padding[-5])]

    return out


def torch_memory(device):
    # Checks and prints GPU memory
    print(f'{torch.cuda.memory_allocated(device)/1024/1024:.2f} MB USED')
    print(f'{torch.cuda.memory_reserved(device)/1024/1024:.2f} MB RESERVED')
    print(f'{torch.cuda.max_memory_allocated(device)/1024/1024:.2f} MB USED MAX')
    print(f'{torch.cuda.max_memory_reserved(device)/1024/1024:.2f} MB RESERVED MAX')
    print('')


device = 0

print(torch.__version__)

original_shape = (1, 32, 24, 256, 256)

# Input to 3D conv
x = torch.randn(original_shape, dtype=torch.float32, device=device)

# Several different padding values
padding_list = [(1, 3, 2, 1, 5, 1),
                (1, 1, 1, 1, 1, 1),
                (0, 0, 0, 0, 0, 0)]

for padding in padding_list:
    # Old circular padding
    y = torch.nn.functional._pad_circular(x, padding)

    # New circular padding
    z = _pad_circular(x, padding)

    assert y.shape == z.shape
    assert torch.all(torch.eq(y, z))

# The old circular padding doesn't work properly for negative values
padding_list = [(0, 0, 0, 0, -1, -1),
                (-1, -1, -1, -1, -1, -1),
                (-2, -3, -4, -12, -5, -1)]

for padding in padding_list:
    # Old
    y = torch.nn.functional._pad_circular(x, padding)
    print(f'Shape does not make sense: {y.shape}')

    # New
    z = _pad_circular(x, padding)
    print(f'Correctly narrows array: {z.shape}')

# Memory tests
y, z = None, None  # Clear memory

# Reset memory
torch.cuda.empty_cache()
torch.cuda.reset_peak_memory_stats()

padding_list = [(1, 3, 2, 1, 5, 1),
                (1, 1, 1, 1, 1, 1),
                (0, 0, 0, 0, 0, 0)]

for padding in padding_list:
    # Old circular padding
    print('Memory before circular padding:')
    torch_memory(device)
    y = torch.nn.functional._pad_circular(x, padding)

    # Check memory used then clear y from memory
    print('Memory used by old circular padding:')
    torch_memory(device)

    y = None
    torch.cuda.empty_cache()
    torch.cuda.reset_peak_memory_stats()

    # New circular padding
    print('Memory before circular padding:')
    torch_memory(device)
    z = _pad_circular(x, padding)

    # Check memory used then clear z from memory
    print('Memory used by new circular padding:')
    torch_memory(device)

    z = None
    torch.cuda.empty_cache()
    torch.cuda.reset_peak_memory_stats()


# Test time

padding_list = [(1, 3, 2, 1, 5, 1),
                (1, 1, 1, 1, 1, 1),
                (0, 0, 0, 0, 0, 0)]

original_shape = (1, 32, 24, 256, 256)
num_repeats = 10

# Test speed on GPU

device = 0

x = torch.randn(original_shape, dtype=torch.float32, device=device)

torch.cuda.synchronize(device)
start_time = time()
for _ in range(num_repeats):
    for padding in padding_list:
        y = F._pad_circular(x, padding)
torch.cuda.synchronize(device)
stop_time = time()
print('Old padding. Time on GPU:', (stop_time - start_time), 's')

torch.cuda.synchronize(device)
start_time = time()
for _ in range(num_repeats):
    for padding in padding_list:
        y = _pad_circular(x, padding)
torch.cuda.synchronize(device)
stop_time = time()
print('New padding. Time on GPU:', (stop_time - start_time), 's')

# Test speed on CPU
device = 'cpu'

x = x.cpu()

start_time = time()
for _ in range(num_repeats):
    for padding in padding_list:
        y = F._pad_circular(x, padding)
stop_time = time()
print('Old padding. Time on CPU:', (stop_time - start_time), 's')

start_time = time()
for _ in range(num_repeats):
    for padding in padding_list:
        y = _pad_circular(x, padding)
stop_time = time()
print('New padding. Time on CPU:', (stop_time - start_time), 's')

```